Resolve nullability warnings
Since there is already a nullability annotation in this header (on the declaration of `-collectionView:layout:scrollViewConfigurationForSection:`), the compiler assumes that the header has been audited for nullability. However, most of the parameters were not marked as audited, which resulted in lots of compiler warnings everywhere this was imported.

With this change, all the warnings go away.

This change does assume that it is invalid to set `defaultScrollViewConfiguration = nil`. If there are cases where that should be allowed, then it would also need to be marked `nullable`, but I don't think that's actually a supported configuration.