Fix NSIndexPath incompatibilites with Swift IndexPath
In Swift, indexPath.section doesn't works the same way as in
NSIndexPath. With NSIndexPath with a single index, indexPath.section
will be equal to that index. This project relied on that to create its
indexPaths. But if that indexPath was used as a Swift IndexPath, it
would crash when accessing .section.

Fixes #13